Stop the overlay edit mode from latching on and stranding the panel
One root cause behind three reports: the panel stayed frozen when the
setting was turned off, the "Настроить оверлей" button appeared to do
nothing, and the panel outlived the call - the screenshot showed it
sitting on the bare desktop, which only happens in edit mode since that
bypasses both the fullscreen gate and teardown.

Edit mode was a plain static bool armed from the editor's constructor,
and Stop() began with "if (_editMode) return". If the editor window
never actually appeared, the flag stayed set forever and Stop() became
a no-op, so nothing could remove the panel again.

- Edit mode is now bound to the editor window instance and is only
  considered active while that window is alive, so it cannot outlive it.
- It is armed from OnShown rather than the constructor: no window on
  screen, no edit mode.
- Stop takes a force flag. Disabling the overlay forces teardown even
  with the editor open - there is nothing to configure on a disabled
  panel. Ending a call without force keeps the editor usable by swapping
  in the sample roster.
- The 700ms tick tears the panel down whenever the setting is off, so
  any missed path self-heals within one tick.
- Stop hides the window before destroying it and disposes it.
- Open() no longer swallows failures: it reports them, clears the static
  handle so the button keeps working, and unwinds edit mode.

Also fixes an OverflowException waiting to happen in the window's
WM_NCHITTEST: LParam was read with ToInt32, which throws on 64-bit once
the packed Y is negative (window above the screen top, second monitor).

Version 2.8.4.1.

+++ PISMO/DarkToolWindow.cs
@@ -123,7 +123,11 @@ namespace PISMO
                 // поэтому границы обозначаем сами.
                 if ((int)m.Result == HTCLIENT)
                 {
-                    int lp = m.LParam.ToInt32();
+                    // ToInt64, а не ToInt32: на 64-битной системе координаты
+                    // упакованы в младшие 32 бита, и при отрицательном Y (окно выше
+                    // верха экрана, второй монитор) ToInt32 бросает OverflowException
+                    // прямо из WndProc.
+                    long lp = m.LParam.ToInt64();
                     var pt = PointToClient(new Point((short)(lp & 0xFFFF), (short)((lp >> 16) & 0xFFFF)));
                     bool l = pt.X <= Grip, r = pt.X >= ClientSize.Width - Grip;
                     bool t = pt.Y <= Grip, b = pt.Y >= ClientSize.Height - Grip;
